Add Common Route Outlet
In React Router DOM, there is the concept of Outlets and Plugs.
Outlets are React components that allows the content of other
routes to be "plugged" into it. This is commonly used to display
a constant set of UI elements that do not change when app routes
are switch.

In this commit, the <Outlet /> component has been replaced with
<RouteOutlet />. The RouteOutlet component is under /src/components/
RouteOutlet/RouteOutlet.tsx. It essentially embeds the original
<Outlet /> inside while displaying a constant header. As routes
are switched between, their content will be "plugged" into the
<Outlet />.