Create Card Component
The two primary files of focus for this commit are /src/components/
Card/Card.tsx and /src/routes/Home/Home.tsx.

If you open, /src/routes/Home/Home.tsx, you will notice it now
renders a <Card /> component. This will be our resuable tool card
for users to click on and view details about the tool.

Feel free to Ctrl + Left Click on the <Card /> component in the
editor to view it's source code! This is the first component in
our source tree that starts to utilize the power of React's props.

Next commit will utilize these props to display something a little
more customized.