Render Fetched Tools as Cards
This is it! The Home route now renders all the fetched tools
as <Card /> components. Notice the use of the map function that
iterates through each tool object and builds an array of <Card />
components. Also note the key prop persent on the <Card />. The key
props should be a unique identifying string (amoung a component's siblings)
that allows React to distinguish between Cards inbetween two renders.
For example, if the <Card /> components happen to change order between
two renders, the key component can help React decide which components
to re-render to the DOM and which ones not to.

For more info on keys, see this link:

https://react.dev/learn/rendering-lists