Prepare View Route for Rendering Tools
Now that we can view tools on the Home route, it's time
to view and modify their details under the View route.

Notice that in /src/routes.tsx there is a placeholder
path for the View route view/:toolId. This will allow
the View component to fetch a particular tool from the
database by its ID.

In /src/routes/View/View.tsx, the hook, useParams is used
to get whatever :toolId parameter is passed to it from the
URL. For now it simply displays the :toolId.