Add Custom useFetchOnce Hook and Fetch Tool
This commit contains a fun new custom hook: /src/hooks/useFetchOnce.
Because it's so common to fetch data in a component and because
components in React can render more than once, this hook triggers
the specified endpoint only once. When the data is recieved the
component using the hook will automatically be re-rendered.

See the Home and View routes for usage. Also notice the URL params
when using the hook in the View route.