Add ToolForm for Viewing and Modifying Tools
There is a bit more to cover in this commit than the previous
ones. I recommend approaching this commit by traversing through
the changes made to the React component tree. You may traverse
through the components in this order (Ctrl + Left Click):

View > ToolForm > UpdateToolInput > HoverInput

Overall, this commit adds the ability for users to modify
the title of a tool by typing into the input on the View
route. The next commit will fill out the rest of the form
with similar inputs.