Build the View Route
This commit contains more grunt work on the View route. It
builds upon the previous in which user input will update
tool data in the database.

If you wish to look through the code a good starting point would
be in the /src/components/ToolForm/ToolForm.tsx which
contains an entire form for viewing and modifying tools and their
associated links.