Add Links to Cards on Home Route
A new controller component has been added under /src/controllers/
CardLink/CardLink.tsx. This is a very simple component wrapping
the Card component with a React Router DOM <Link></Link>.

Notice in the Home route how all the Card components are now
CardLinks. Only one prop was added and that was the to prop
which specifies the proper route to view the element based on
the tool's id.