Add Fixed Panel Component for Fixed Positioning
The next step in the app is to make a button for adding a new tool.
The button will create a new blank tool in the database and then
link the user over to the proper View route for modifying that
tool.

This first step contains a component called FixedPanel that allows
developers to display fixed content on the screen. This will be used
eventually on the Home and View routes for buttons that create and
delete tools.