Add Fixed "Add Tool" Button on Home Route
The Home route now contains a circle button that will be used
for creating a new tool and bringing users to the proper View
route for that tool.

This commit only focuses on the styling of the button which can be
found as class .circleButton under /src/styles/global-styles.css.
Apply this style to any element you want to appear as a circular
button. Button specific styles like color are to be defined on
a specific element.

Next commit will create a controller for this button to both trigger
CreateEmptyTool.php and navigate to the proper View route.