Minor Changes to CSS for Mobile Layouts
This commit is a good example of using the relative CSS units,
rem and em to your advantage. Because the project was designed
with those units, the majority of the change here is occuring in
/src/components/RouteOutlet/RouteOutlet.module.css. There are
two media query changing the font size for two breakpoints. Notice
how the rest of the app simply responds when the font-size of a
high-up ancentor element changes its font-size. If elements further
down in the DOM used px instead of em, then they wouldn't
resize.